Restructuring bigquery and spark consumers (#73)
* Log metric values as info message when MaxBatchSizeWait is used

* Restructuring bigquery and spark consumers

+++ debezium-server-batch-sinks/src/test/java/io/debezium/server/batch/common/DynamicBatchSizeWaitTest.java
@@ -25,47 +25,47 @@ class DynamicBatchSizeWaitTest {
   DynamicBatchSizeWait waitBatchSize;
 
   @ConfigProperty(name = "debezium.source.poll.interval.ms", defaultValue = "1000")
-  Integer pollIntervalMs;
+  long pollIntervalMs;
 
   @Test
   void shouldIncreaseSleepMs() {
     DynamicBatchSizeWait dynamicSleep = waitBatchSize;
     // if its consuming small batch sizes, the sleep delay should increase to adjust batch size
     // sleep size should increase and stay at max (pollIntervalMs)
-    int sleep = 0;
-    sleep = dynamicSleep.getWaitMs(3);
+    long sleep = 0;
+    sleep = dynamicSleep.getWaitMs(3L);
     Assertions.assertTrue(sleep < pollIntervalMs);
-    sleep = dynamicSleep.getWaitMs(2);
+    sleep = dynamicSleep.getWaitMs(2L);
     Assertions.assertTrue(sleep <= pollIntervalMs);
-    sleep = dynamicSleep.getWaitMs(1);
-    Assertions.assertEquals((Integer) sleep, pollIntervalMs);
-    sleep = dynamicSleep.getWaitMs(1);
-    Assertions.assertEquals((Integer) sleep, pollIntervalMs);
-    sleep = dynamicSleep.getWaitMs(1);
-    Assertions.assertEquals((Integer) sleep, pollIntervalMs);
+    sleep = dynamicSleep.getWaitMs(1L);
+    Assertions.assertEquals(sleep, pollIntervalMs);
+    sleep = dynamicSleep.getWaitMs(1L);
+    Assertions.assertEquals(sleep, pollIntervalMs);
+    sleep = dynamicSleep.getWaitMs(1L);
+    Assertions.assertEquals(sleep, pollIntervalMs);
   }
 
   @Test
   void shouldDecreaseSleepMs() {
     DynamicBatchSizeWait dynamicSleep = waitBatchSize;
     // if its consuming large batch sizes, the sleep delay should decrease
-    dynamicSleep.getWaitMs(3);
-    dynamicSleep.getWaitMs(2);
-    dynamicSleep.getWaitMs(1);
+    dynamicSleep.getWaitMs(3L);
+    dynamicSleep.getWaitMs(2L);
+    dynamicSleep.getWaitMs(1L);
     // start test
     // max batch size = debezium.source.max.batch.size = 100
-    int sleep1 = dynamicSleep.getWaitMs(120);
-    int sleep2 = dynamicSleep.getWaitMs(120);
+    long sleep1 = dynamicSleep.getWaitMs(120L);
+    long sleep2 = dynamicSleep.getWaitMs(120L);
     Assertions.assertTrue(sleep2 <= sleep1);
-    int sleep3 = dynamicSleep.getWaitMs(120);
+    long sleep3 = dynamicSleep.getWaitMs(120L);
     Assertions.assertTrue(sleep3 <= sleep2);
-    int sleep4 = dynamicSleep.getWaitMs(120);
+    long sleep4 = dynamicSleep.getWaitMs(120L);
     Assertions.assertTrue(sleep4 <= sleep3);
-    dynamicSleep.getWaitMs(120);
-    dynamicSleep.getWaitMs(120);
-    dynamicSleep.getWaitMs(120);
-    dynamicSleep.getWaitMs(120);
-    Assertions.assertTrue(dynamicSleep.getWaitMs(120) <= 100);
+    dynamicSleep.getWaitMs(120L);
+    dynamicSleep.getWaitMs(120L);
+    dynamicSleep.getWaitMs(120L);
+    dynamicSleep.getWaitMs(120L);
+    Assertions.assertTrue(dynamicSleep.getWaitMs(120L) <= 100L);
   }
 
 }
